SelectedFatmanManager, apply addUser Fatman updated when purchase avata

+++ fatman/app/src/main/java/com/project/fat/data/dto/Fatman.kt
@@ -10,5 +10,5 @@ data class FatmanElement (
     val name: String,
     val fatmanImageUrl: String,
     val fatmanId: Long,
-    val fatmanCost: Int
+    val fatmanCost: Long
 )
